New Faces panel
The new panel permits to define the image topic and tracked
faces topic; this been done, it is possible to visualize
in the new panel the images coming from the camera
and, over these, the bounding box of the detected face (if any).
At this point, the panel is capable to display only one face.
Work in Progress.